refactor: remove unused context and effect components, streamline WeaverseHydrogen structure
- Deleted context.ts and Effect.tsx files to eliminate unused code.
- Refactored WeaverseHydrogenRoot to integrate new utility functions and improve component registration.
- Consolidated image placeholders into a single utility file for better organization.
- Enhanced WeaverseHydrogen and WeaverseHydrogenItem classes for improved clarity and functionality.
