+++ packages/hydrogen/src/index.ts
@@ -1,7 +1,6 @@
+import { getSelectedProductOptions as hydrogen_getSelectedProductOptions } from '@shopify/hydrogen'
+import type { SelectedOptionInput } from '@shopify/hydrogen/storefront-api-types'
 import {
-  type PlatformTypeEnum,
-  Weaverse,
-  WeaverseItemStore,
   isBrowser,
   isIframe,
   useChildInstances,
@@ -9,70 +8,85 @@ import {
   useParentInstance,
   useWeaverse,
 } from '@weaverse/react'
-import type {
-  HydrogenComponentData,
-  HydrogenElement,
-  HydrogenPageData,
-  WeaverseHydrogenParams,
-  WeaverseInternal,
-  WeaverseLoaderRequestInfo,
-} from './types'
-import { generateDataFromSchema } from './utils'
 
-export * from './WeaverseHydrogenRoot'
-export { useThemeSettings } from './hooks/use-theme-settings'
-export * from './types'
-export * from './utils'
-export * from './weaverse-client'
-export * from './wrappers'
-export * from './placeholders'
+export function getSelectedProductOptions(
+  request: Request,
+): SelectedOptionInput[] {
+  let options = hydrogen_getSelectedProductOptions(request)
+  return options.filter(
+    ({ name }) => name !== 'isDesignMode' && !name.startsWith('weaverse'),
+  )
+}
+
+export const IMAGES_PLACEHOLDERS = {
+  logo_white:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-logo-w-600x200.svg',
+  logo_black:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-logo-k-600x200.svg',
+  image:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-image-1024x1024.svg',
+  banner_1:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-lifestyle-1-1800x900.svg',
+  banner_2:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-lifestyle-2-1800x900.svg',
+  collection_1:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-collection-hats-1024x1024.svg',
+  collection_2:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-collection-shoes-1024x1024.svg',
+  collection_3:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-collection-glasses-1024x1024.svg',
+  collection_4:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-collection-lamps-1024x1024.svg',
+  collection_5:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-collection-bags-1024x1024.svg',
+  collection_6:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-collection-watches-1024x1024.svg',
+  product_1:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-watch-3-1024x1024.svg',
+  product_2:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-bag-2-1024x1024.svg',
+  product_3:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-shoe-1-1024x1024.svg',
+  product_4:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-lamp-2-1024x1024.svg',
+  product_5:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-glasses-1-1024x1024.svg',
+  product_6:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-lamp-1-1024x1024.svg',
+  product_7:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-glasses-2-1024x1024.svg',
+  product_8:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-hat-3-1024x1024.svg',
+  product_9:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-bag-1-1024x1024.svg',
+  product_10:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-shoe-2-1024x1024.svg',
+  product_11:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-hat-2-1024x1024.svg',
+  product_12:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-glasses-3-1024x1024.svg',
+  product_13:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-shoe-3-1024x1024.svg',
+  product_14:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-bag-3-1024x1024.svg',
+  product_15:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-hat-1-1024x1024.svg',
+  product_16:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-lamp-3-1024x1024.svg',
+  product_17:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-watch-1-1024x1024.svg',
+  product_18:
+    'https://cdn.shopify.com/s/files/1/0838/0052/3057/files/fpo-product-watch-2-1024x1024.svg',
+}
+
 export {
+  isBrowser,
+  isIframe,
   useChildInstances,
   useItemInstance,
   useParentInstance,
   useWeaverse,
-  isBrowser,
-  isIframe,
 }
-
-export class WeaverseHydrogen extends Weaverse {
-  platformType: PlatformTypeEnum = 'shopify-hydrogen'
-  pageId: string
-  internal: Partial<WeaverseInternal>
-  requestInfo: WeaverseLoaderRequestInfo
-  declare ItemConstructor: typeof WeaverseHydrogenItem
-  declare data: HydrogenPageData
-  static itemInstances: Map<string, WeaverseHydrogenItem>
-  static elementRegistry: Map<string, HydrogenElement>
-
-  constructor(params: WeaverseHydrogenParams) {
-    let { internal, pageId, requestInfo, ...coreParams } = params
-    super({ ...coreParams })
-    this.internal = internal
-    this.pageId = pageId
-    this.requestInfo = requestInfo
-  }
-}
-
-export class WeaverseHydrogenItem extends WeaverseItemStore {
-  declare weaverse: WeaverseHydrogen
-
-  constructor(initialData: HydrogenComponentData, weaverse: WeaverseHydrogen) {
-    super(initialData, weaverse)
-    let { data, ...rest } = initialData
-    Object.assign(this._store, this._schemaData, data, rest)
-  }
-
-  get Element(): HydrogenElement {
-    return super.Element
-  }
-
-  get _schemaData() {
-    if (this.Element?.schema) {
-      return generateDataFromSchema(this.Element.schema)
-    }
-    return {}
-  }
-}
-
-Weaverse.ItemConstructor = WeaverseHydrogenItem
+export * from './types'
+export * from './weaverse-client'
+export * from './WeaverseHydrogenRoot'
